requirements.txt
PyQt6>=6.5.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
tortoise-orm>=0.20.0
fpdf2>=2.7.0
nest-asyncio>=1.5.0
python-dotenv>=1.0.0
